day 11 part 3

+++ Day11/Day11.h
@@ -11,7 +11,7 @@ public:
 
     static int Part2();
 
-    static int Part3();
+    static long long Part3();
 };
 
 #endif //DAY11_H
